trying to render two differnt worlds.

+++ code/server/sv_world.c
@@ -61,16 +61,8 @@ them, which prevents having to deal with multiple fragments of a single entity.
 ===============================================================================
 */
 
-typedef struct worldSector_s {
-	int		axis;		// -1 = leaf node
-	float	dist;
-	struct worldSector_s	*children[2];
-	svEntity_t	*entities;
-} worldSector_t;
-
-#define	AREA_DEPTH	4
-#define	AREA_NODES	64
-
+worldSector_t	serverWorlds[10][AREA_NODES];
+int 			numServerWorlds = 0;
 worldSector_t	sv_worldSectors[AREA_NODES];
 int			sv_numworldSectors;
 
@@ -148,10 +140,9 @@ void SV_ClearWorld( void ) {
 	clipHandle_t	h;
 	vec3_t			mins, maxs;
 
-	if(sv.state == SS_GAME) {
-		CM_SwitchMap(0);
-	}
-
+if(numServerWorlds > 0) {
+	memcpy(serverWorlds[numServerWorlds-1], sv_worldSectors, sizeof(sv_worldSectors));
+}
 	Com_Memset( sv_worldSectors, 0, sizeof(sv_worldSectors) );
 	sv_numworldSectors = 0;
 
@@ -159,6 +150,7 @@ void SV_ClearWorld( void ) {
 	h = CM_InlineModel( 0 );
 	CM_ModelBounds( h, mins, maxs );
 	SV_CreateworldSector( 0, mins, maxs );
+	numServerWorlds++;
 }
 
 
@@ -174,6 +166,9 @@ void SV_UnlinkEntity( sharedEntity_t *gEnt ) {
 	worldSector_t	*ws;
 
 	ent = SV_SvEntityForGentity( gEnt );
+	if(!ent) {
+		return;
+	}
 
 	gEnt->r.linked = qfalse;
 
@@ -218,6 +213,9 @@ void SV_LinkEntity( sharedEntity_t *gEnt ) {
 	svEntity_t	*ent;
 
 	ent = SV_SvEntityForGentity( gEnt );
+	if(!ent) {
+		return;
+	}
 
 	if ( ent->worldSector ) {
 		SV_UnlinkEntity( gEnt );	// unlink from old position
